Типы данных.Переменные.Неявное преобразование.Чиcло pi в java.

+++ lesson1_2/levelA/a2/Run.java
@@ -6,7 +6,7 @@ package lesson1_2.levelA.a2;
 public class Run {
     public static void main(String[] args) {
         byte varByte = 10;
-        short varShort = varByte;
+        short varShort = (short) varByte;
         System.out.println(varByte);
         System.out.println(varShort);
     }
